Refactor signatures for transaction forwarding, fix tests
