+++ services/transactionpool/forward_transactions.go
@@ -9,8 +9,32 @@ import (
 	"github.com/orbs-network/orbs-spec/types/go/protocol"
 	"github.com/orbs-network/orbs-spec/types/go/protocol/gossipmessages"
 	"github.com/orbs-network/orbs-spec/types/go/services/gossiptopics"
+	"github.com/orbs-network/orbs-spec/types/go/services/handlers"
+	"github.com/pkg/errors"
 	"time"
 )
+
+func (s *service) RegisterTransactionResultsHandler(handler handlers.TransactionResultsHandler) {
+	s.transactionResultsHandlers = append(s.transactionResultsHandlers, handler)
+}
+
+func (s *service) HandleForwardedTransactions(input *gossiptopics.ForwardedTransactionsInput) (*gossiptopics.EmptyOutput, error) {
+	sender := input.Message.Sender
+	oneBigHash, _ := HashTransactions(input.Message.SignedTransactions)
+
+	if !signature.VerifyEd25519(sender.SenderPublicKey(), oneBigHash, sender.Signature()) {
+		return nil, errors.Errorf("invalid signature in relay message from sender %s", sender.SenderPublicKey())
+	}
+
+	for _, tx := range input.Message.SignedTransactions {
+		txHash := digest.CalcTxHash(tx.Transaction())
+		s.logger.Info("adding forwarded transaction to the pool", log.String("flow", "checkpoint"), log.Stringable("transaction", tx), log.Stringable("txHash", txHash))
+		if _, err := s.pendingPool.add(tx, sender.SenderPublicKey()); err != nil {
+			s.logger.Error("error adding forwarded transaction to pending pool", log.Error(err), log.Stringable("transaction", tx), log.Stringable("txHash", txHash))
+		}
+	}
+	return nil, nil
+}
 
 func (s *service) startForwardingProcess(ctx context.Context) {
 	go func() {
@@ -32,7 +56,7 @@ func (s *service) drainQueueAndForwardTransactions() {
 		return
 	}
 
-	oneBigHash, hashes := hashTransactions(txs)
+	oneBigHash, hashes := HashTransactions(txs)
 
 	sig, err := signature.SignEd25519(s.config.NodePrivateKey(), oneBigHash)
 	if err != nil {
@@ -52,9 +76,9 @@ func (s *service) drainQueueAndForwardTransactions() {
 
 	for _, hash := range hashes {
 		if err != nil {
-			s.logger.Info("failed forwarding transaction via gossip", log.Error(err), log.Stringable("txHash", hash))
+			s.logger.Info("failed forwarding transaction via gossip", log.Error(err), log.String("flow", "checkpoint"), log.Stringable("txHash", hash))
 		} else {
-			s.logger.Info("forwarded transaction via gossip", log.Stringable("txHash", hash))
+			s.logger.Info("forwarded transaction via gossip", log.String("flow", "checkpoint"), log.Stringable("txHash", hash))
 		}
 	}
 }
@@ -67,7 +91,7 @@ func (s *service) drainQueue() []*protocol.SignedTransaction {
 	return txs
 }
 
-func hashTransactions(txs []*protocol.SignedTransaction) (oneBigHash []byte, hashes []primitives.Sha256) {
+func HashTransactions(txs []*protocol.SignedTransaction) (oneBigHash []byte, hashes []primitives.Sha256) {
 	for _, tx := range txs {
 		hash := digest.CalcTxHash(tx.Transaction())
